backend/app/db/connection.py
import psycopg2
from app.core.config import settings

def get_db():
    return psycopg2.connect(
        host=settings.DB_HOST,
        database=settings.DB_NAME,
        user=settings.DB_USER,
        password=settings.DB_PASSWORD
    )